requirements.txt
web3>=6.11.0
eth-account>=0.10.0
python-dotenv>=1.0.0
eth-abi>=4.2.0
websockets>=12.0
python-socks[asyncio]>=2.4.4
